Refactor: 리뷰 받은 코드 수정
- JwtFilter.jwtExceptionHandler -> JwtFilter.handleJwtException 리네임
- SecurityConfig 클래스 swagger 패키지에서 분리
